requirements.txt
tensorflow
keras
streamlit
Pillow
numpy
opencv-python
matplotlib
seaborn
pandas
scikit-learn
